// HciMaliProjekat/src/app/model/country.model.ts
export interface Country {
  name: String;
  flag: String;
  capitalCity: String;
  currencies: String[];
  population: Number;
  continents: String[];
  region: String;
  subregion: String;
  latitude: Number;
  longitude: Number;
  languages: String[];
  map: String;
  crest: String;
  timeZones: String[];
  postalCode: String;
}